Switch back to xml
I have tried using markdown and have found it just to be a pain as I
don't know how to do too many things - such as having both comments and
keeping old versions of things (i.e. xml comments).  This is the same as
the last version of draft-schaad-cose.md.  The name change was necessary
to make the IETF data tracker happy.